Split font related css property parsing out into its own file https://bugs.webkit.org/show_bug.cgi?id=278034
Reviewed by Darin Adler.

Split font related CSS property parsing into its own file, consolidating
parts from CSSPropertyParserHelpers.h/cpp and CSSPropertyParserWorkerSafe.h/cpp.

Do a minor cleanup of names, but otherwise, the code has been copied as is.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSFontFaceSet.cpp:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/FontFace.cpp:
* Source/WebCore/css/ShorthandSerializer.cpp:
* Source/WebCore/css/StyleProperties.cpp:
* Source/WebCore/css/parser/CSSParserImpl.cpp:
* Source/WebCore/css/parser/CSSPropertyParser.cpp:
* Source/WebCore/css/parser/CSSPropertyParser.h:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Font.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Font.h: Added.
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
* Source/WebCore/css/parser/CSSPropertyParserWorkerSafe.cpp:
* Source/WebCore/css/parser/CSSPropertyParserWorkerSafe.h:
* Source/WebCore/css/parser/CSSSupportsParser.cpp:
* Source/WebCore/css/process-css-properties.py:
* Source/WebCore/editing/EditingStyle.cpp:
* Source/WebCore/html/canvas/CanvasRenderingContext2D.cpp:
* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.cpp:
* Source/WebCore/html/canvas/OffscreenCanvasRenderingContext2D.cpp:
* Source/WebCore/style/StyleBuilderConverter.h:
* Source/WebCore/style/StyleBuilderCustom.h:
* Source/WebCore/style/StyleResolveForFontRaw.cpp:

Canonical link: https://commits.webkit.org/282527@main
