+++ Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp
@@ -47,7 +47,6 @@
 #include "CSSPropertyParserConsumer+PercentDefinitions.h"
 #include "CSSPropertyParserConsumer+Position.h"
 #include "CSSPropertyParserConsumer+Primitives.h"
-#include "CSSPropertyParserConsumer+RawResolver.h"
 #include "CSSPropertyParserConsumer+String.h"
 #include "CSSPropertyParserConsumer+Time.h"
 #include "CSSPropertyParserConsumer+URL.h"
@@ -65,11 +64,6 @@
 #include "CSSCounterValue.h"
 #include "CSSCursorImageValue.h"
 #include "CSSCustomPropertyValue.h"
-#include "CSSFontFaceSrcValue.h"
-#include "CSSFontPaletteValuesOverrideColorsValue.h"
-#include "CSSFontVariantAlternatesValue.h"
-#include "CSSFontVariantLigaturesParser.h"
-#include "CSSFontVariantNumericParser.h"
 #include "CSSFunctionValue.h"
 #include "CSSGridAutoRepeatValue.h"
 #include "CSSGridIntegerRepeatValue.h"
@@ -98,8 +92,6 @@
 #include "CSSViewValue.h"
 #include "CalculationCategory.h"
 #include "ColorInterpolation.h"
-#include "FontCustomPlatformData.h"
-#include "FontFace.h"
 #include "LengthPoint.h"
 #include "Logging.h"
 #include "RenderStyleConstants.h"
@@ -108,66 +100,12 @@
 #include "ScriptExecutionContext.h"
 #include "StyleColor.h"
 #include "TimingFunction.h"
-#include "WebKitFontFamilyNames.h"
 #include <wtf/SortedArrayMap.h>
 #include <wtf/text/TextStream.h>
 
 namespace WebCore {
 
 namespace CSSPropertyParserHelpers {
-
-// MARK: - Non-primitive consumers.
-
-static std::optional<double> consumeFontWeightNumberRaw(CSSParserTokenRange& range)
-{
-#if !ENABLE(VARIATION_FONTS)
-    auto isIntegerAndDivisibleBy100 = [](double value) {
-        ASSERT(value >= 1);
-        ASSERT(value <= 1000);
-        return static_cast<int>(value / 100) * 100 == value;
-    };
-#endif
-
-    auto& token = range.peek();
-    switch (token.type()) {
-    case FunctionToken: {
-        // "[For calc()], the used value resulting from an expression must be clamped to the range allowed in the target context."
-        auto unevaluatedCalc = NumberKnownTokenTypeFunctionConsumer::consume(range, { }, { });
-        if (!unevaluatedCalc)
-            return std::nullopt;
-        auto result = RawResolver<NumberRaw>::resolve(*unevaluatedCalc, { }, { });
-        if (!result)
-            return std::nullopt;
-#if !ENABLE(VARIATION_FONTS)
-        if (!(result->value >= 1 && result->value <= 1000) || !isIntegerAndDivisibleBy100(result->value))
-            return std::nullopt;
-#endif
-        return std::clamp(result->value, 1.0, 1000.0);
-    }
-
-    case NumberToken: {
-        auto result = token.numericValue();
-        if (!(result >= 1 && result <= 1000))
-            return std::nullopt;
-#if !ENABLE(VARIATION_FONTS)
-        if (token.numericValueType() != IntegerValueType || !isIntegerAndDivisibleBy100(result))
-            return std::nullopt;
-#endif
-        range.consumeIncludingWhitespace();
-        return result;
-    }
-
-    default:
-        return std::nullopt;
-    }
-}
-
-RefPtr<CSSPrimitiveValue> consumeFontWeightNumber(CSSParserTokenRange& range)
-{
-    if (auto result = consumeFontWeightNumberRaw(range))
-        return CSSPrimitiveValue::create(*result);
-    return nullptr;
-}
 
 static bool isPixelFilterFunction(CSSValueID filterFunction)
 {
@@ -406,266 +344,6 @@ RefPtr<CSSPrimitiveValue> consumeSingleContainerName(CSSParserTokenRange& range)
     }
 }
 
-static std::optional<FontWeightRaw> consumeFontWeightRaw(CSSParserTokenRange& range)
-{
-    if (auto result = consumeIdentRaw<CSSValueNormal, CSSValueBold, CSSValueBolder, CSSValueLighter>(range))
-        return { *result };
-    if (auto result = consumeFontWeightNumberRaw(range))
-        return { *result };
-    return std::nullopt;
-}
-
-std::optional<CSSValueID> consumeFontStretchKeywordValueRaw(CSSParserTokenRange& range)
-{
-    return consumeIdentRaw<CSSValueUltraCondensed, CSSValueExtraCondensed, CSSValueCondensed, CSSValueSemiCondensed, CSSValueNormal, CSSValueSemiExpanded, CSSValueExpanded, CSSValueExtraExpanded, CSSValueUltraExpanded>(range);
-}
-
-static std::optional<FontStyleRaw> consumeFontStyleRaw(CSSParserTokenRange& range, CSSParserMode parserMode)
-{
-#if ENABLE(VARIATION_FONTS)
-    CSSParserTokenRange rangeBeforeKeyword = range;
-#endif
-
-    auto keyword = consumeIdentRaw<CSSValueNormal, CSSValueItalic, CSSValueOblique>(range);
-    if (!keyword)
-        return std::nullopt;
-
-#if ENABLE(VARIATION_FONTS)
-    if (*keyword == CSSValueOblique && !range.atEnd()) {
-        // FIXME: This angle does specify that unitless 0 is allowed - see https://drafts.csswg.org/css-fonts-4/#valdef-font-style-oblique-angle
-        if (auto angle = consumeAngleRaw(range, parserMode, UnitlessQuirk::Forbid, UnitlessZeroQuirk::Allow)) {
-            if (isFontStyleAngleInRange(CSSPrimitiveValue::computeDegrees(angle->type, angle->value)))
-                return { { CSSValueOblique, WTFMove(angle) } };
-            // Must not consume anything on error.
-            range = rangeBeforeKeyword;
-            return std::nullopt;
-        }
-    }
-#else
-    UNUSED_PARAM(parserMode);
-#endif
-
-    return { { *keyword, std::nullopt } };
-}
-
-AtomString concatenateFamilyName(CSSParserTokenRange& range)
-{
-    StringBuilder builder;
-    bool addedSpace = false;
-    const CSSParserToken& firstToken = range.peek();
-    while (range.peek().type() == IdentToken) {
-        if (!builder.isEmpty()) {
-            builder.append(' ');
-            addedSpace = true;
-        }
-        builder.append(range.consumeIncludingWhitespace().value());
-    }
-    // <family-name> can't contain unquoted generic families
-    if (!addedSpace && (!isValidCustomIdentifier(firstToken.id()) || !genericFontFamily(firstToken.id()).isNull()))
-        return nullAtom();
-    return builder.toAtomString();
-}
-
-AtomString consumeFamilyNameRaw(CSSParserTokenRange& range)
-{
-    if (range.peek().type() == StringToken)
-        return range.consumeIncludingWhitespace().value().toAtomString();
-    if (range.peek().type() != IdentToken)
-        return nullAtom();
-    return concatenateFamilyName(range);
-}
-
-Vector<AtomString> consumeFamilyNameListRaw(CSSParserTokenRange& range)
-{
-    Vector<AtomString> result;
-    do {
-        auto name = consumeFamilyNameRaw(range);
-        if (name.isNull())
-            return { };
-
-        result.append(name);
-    } while (consumeCommaIncludingWhitespace(range));
-    return result;
-}
-
-static std::optional<CSSValueID> consumeGenericFamilyRaw(CSSParserTokenRange& range)
-{
-    return consumeIdentRaw<CSSValueSerif, CSSValueSansSerif, CSSValueCursive, CSSValueFantasy, CSSValueMonospace, CSSValueWebkitBody, CSSValueWebkitPictograph, CSSValueSystemUi>(range);
-}
-
-static RefPtr<CSSValue> consumeGenericFamily(CSSParserTokenRange& range)
-{
-    if (auto familyName = consumeGenericFamilyRaw(range)) {
-        // FIXME: Remove special case for system-ui.
-        if (*familyName == CSSValueSystemUi)
-            return CSSValuePool::singleton().createFontFamilyValue(nameString(*familyName));
-        return CSSPrimitiveValue::create(*familyName);
-    }
-    return nullptr;
-}
-
-static std::optional<Vector<FontFamilyRaw>> consumeFontFamilyRaw(CSSParserTokenRange& range)
-{
-    Vector<FontFamilyRaw> list;
-    do {
-        if (auto ident = consumeGenericFamilyRaw(range))
-            list.append({ *ident });
-        else {
-            auto familyName = consumeFamilyNameRaw(range);
-            if (familyName.isNull())
-                return std::nullopt;
-            list.append({ familyName });
-        }
-    } while (consumeCommaIncludingWhitespace(range));
-    return list;
-}
-
-RefPtr<CSSValue> consumeFontSizeAdjust(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueNone || range.peek().id() == CSSValueFromFont)
-        return consumeIdent(range);
-
-    auto metric = consumeIdent<CSSValueExHeight, CSSValueCapHeight, CSSValueChWidth, CSSValueIcWidth, CSSValueIcHeight>(range);
-    auto value = consumeNumber(range, ValueRange::NonNegative);
-    if (!value)
-        value = consumeIdent<CSSValueFromFont>(range);
-
-    if (!value || !metric || metric->valueID() == CSSValueExHeight)
-        return value;
-
-    return CSSValuePair::create(metric.releaseNonNull(), value.releaseNonNull());
-}
-
-static std::optional<FontSizeRaw> consumeFontSizeRaw(CSSParserTokenRange& range, CSSParserMode parserMode)
-{
-    // -webkit-xxx-large is a parse-time alias.
-    if (range.peek().id() == CSSValueWebkitXxxLarge) {
-        if (auto ident = consumeIdentRaw(range); ident && ident == CSSValueWebkitXxxLarge)
-            return { CSSValueXxxLarge };
-        return std::nullopt;
-    }
-
-    if (range.peek().id() >= CSSValueXxSmall && range.peek().id() <= CSSValueLarger) {
-        if (auto ident = consumeIdentRaw(range))
-            return { *ident };
-        return std::nullopt;
-    }
-
-    if (auto result = consumeLengthOrPercentRaw(range, parserMode))
-        return { *result };
-
-    return std::nullopt;
-}
-
-static std::optional<LineHeightRaw> consumeLineHeightRaw(CSSParserTokenRange& range, CSSParserMode parserMode)
-{
-    if (range.peek().id() == CSSValueNormal) {
-        if (auto ident = consumeIdentRaw(range))
-            return { *ident };
-        return std::nullopt;
-    }
-
-    if (auto number = consumeNumberRaw(range, ValueRange::NonNegative))
-        return { number->value };
-
-    if (auto lengthOrPercent = consumeLengthOrPercentRaw(range, parserMode))
-        return { *lengthOrPercent };
-
-    return std::nullopt;
-}
-
-std::optional<FontRaw> consumeFontRaw(CSSParserTokenRange& range, CSSParserMode parserMode)
-{
-    FontRaw result;
-
-    for (unsigned i = 0; i < 4 && !range.atEnd(); ++i) {
-        if (consumeIdentRaw<CSSValueNormal>(range))
-            continue;
-        if (!result.style && (result.style = consumeFontStyleRaw(range, parserMode)))
-            continue;
-        if (!result.variantCaps && (result.variantCaps = consumeIdentRaw<CSSValueSmallCaps>(range)))
-            continue;
-        if (!result.weight && (result.weight = consumeFontWeightRaw(range)))
-            continue;
-        if (!result.stretch && (result.stretch = consumeFontStretchKeywordValueRaw(range)))
-            continue;
-        break;
-    }
-
-    if (range.atEnd())
-        return std::nullopt;
-
-    // Now a font size _must_ come.
-    if (auto size = consumeFontSizeRaw(range, parserMode))
-        result.size = *size;
-    else
-        return std::nullopt;
-
-    if (range.atEnd())
-        return std::nullopt;
-
-    if (consumeSlashIncludingWhitespace(range)) {
-        if (!(result.lineHeight = consumeLineHeightRaw(range, parserMode)))
-            return std::nullopt;
-    }
-
-    // Font family must come now.
-    if (auto family = consumeFontFamilyRaw(range))
-        result.family = *family;
-    else
-        return std::nullopt;
-
-    if (!range.atEnd())
-        return std::nullopt;
-
-    return result;
-}
-
-const AtomString& genericFontFamily(CSSValueID ident)
-{
-    switch (ident) {
-    case CSSValueSerif:
-        return WebKitFontFamilyNames::serifFamily.get();
-    case CSSValueSansSerif:
-        return WebKitFontFamilyNames::sansSerifFamily.get();
-    case CSSValueCursive:
-        return WebKitFontFamilyNames::cursiveFamily.get();
-    case CSSValueFantasy:
-        return WebKitFontFamilyNames::fantasyFamily.get();
-    case CSSValueMonospace:
-        return WebKitFontFamilyNames::monospaceFamily.get();
-    case CSSValueWebkitPictograph:
-        return WebKitFontFamilyNames::pictographFamily.get();
-    case CSSValueSystemUi:
-        return WebKitFontFamilyNames::systemUiFamily.get();
-    default:
-        return nullAtom();
-    }
-}
-
-WebKitFontFamilyNames::FamilyNamesIndex genericFontFamilyIndex(CSSValueID ident)
-{
-    switch (ident) {
-    case CSSValueSerif:
-        return WebKitFontFamilyNames::FamilyNamesIndex::SerifFamily;
-    case CSSValueSansSerif:
-        return WebKitFontFamilyNames::FamilyNamesIndex::SansSerifFamily;
-    case CSSValueCursive:
-        return WebKitFontFamilyNames::FamilyNamesIndex::CursiveFamily;
-    case CSSValueFantasy:
-        return WebKitFontFamilyNames::FamilyNamesIndex::FantasyFamily;
-    case CSSValueMonospace:
-        return WebKitFontFamilyNames::FamilyNamesIndex::MonospaceFamily;
-    case CSSValueWebkitPictograph:
-        return WebKitFontFamilyNames::FamilyNamesIndex::PictographFamily;
-    case CSSValueSystemUi:
-        return WebKitFontFamilyNames::FamilyNamesIndex::SystemUiFamily;
-    default:
-        ASSERT_NOT_REACHED();
-        return WebKitFontFamilyNames::FamilyNamesIndex::StandardFamily;
-    }
-}
-
 static RefPtr<CSSValueList> consumeAspectRatioValue(CSSParserTokenRange& range)
 {
     auto leftValue = consumeNumber(range, ValueRange::NonNegative);
@@ -887,282 +565,6 @@ RefPtr<CSSValue> consumeQuotes(CSSParserTokenRange& range)
     if (values.size() && !(values.size() % 2))
         return CSSValueList::createSpaceSeparated(WTFMove(values));
     return nullptr;
-}
-
-RefPtr<CSSValue> consumeFontVariantLigatures(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueNormal || range.peek().id() == CSSValueNone)
-        return consumeIdent(range);
-
-    CSSFontVariantLigaturesParser ligaturesParser;
-    do {
-        if (ligaturesParser.consumeLigature(range) != CSSFontVariantLigaturesParser::ParseResult::ConsumedValue)
-            return nullptr;
-    } while (!range.atEnd());
-
-    return ligaturesParser.finalizeValue();
-}
-
-RefPtr<CSSValue> consumeFontVariantEastAsian(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueNormal)
-        return consumeIdent(range);
-
-    std::optional<FontVariantEastAsianVariant> variant;
-    std::optional<FontVariantEastAsianWidth> width;
-    std::optional<FontVariantEastAsianRuby> ruby;
-
-    auto parseSomethingWithoutError = [&range, &variant, &width, &ruby] () {
-        bool hasParsedSomething = false;
-
-        while (true) {
-            if (range.peek().type() != IdentToken)
-                return hasParsedSomething;
-
-            switch (range.peek().id()) {
-            case CSSValueJis78:
-                if (variant)
-                    return false;
-                variant = FontVariantEastAsianVariant::Jis78;
-                break;
-            case CSSValueJis83:
-                if (variant)
-                    return false;
-                variant = FontVariantEastAsianVariant::Jis83;
-                break;
-            case CSSValueJis90:
-                if (variant)
-                    return false;
-                variant = FontVariantEastAsianVariant::Jis90;
-                break;
-            case CSSValueJis04:
-                if (variant)
-                    return false;
-                variant = FontVariantEastAsianVariant::Jis04;
-                break;
-            case CSSValueSimplified:
-                if (variant)
-                    return false;
-                variant = FontVariantEastAsianVariant::Simplified;
-                break;
-            case CSSValueTraditional:
-                if (variant)
-                    return false;
-                variant = FontVariantEastAsianVariant::Traditional;
-                break;
-            case CSSValueFullWidth:
-                if (width)
-                    return false;
-                width = FontVariantEastAsianWidth::Full;
-                break;
-            case CSSValueProportionalWidth:
-                if (width)
-                    return false;
-                width = FontVariantEastAsianWidth::Proportional;
-                break;
-            case CSSValueRuby:
-                if (ruby)
-                    return false;
-                ruby = FontVariantEastAsianRuby::Yes;
-                break;
-            default:
-                return hasParsedSomething;
-            }
-
-            range.consumeIncludingWhitespace();
-            hasParsedSomething = true;
-        }
-    };
-
-    if (!parseSomethingWithoutError())
-        return nullptr;
-
-    CSSValueListBuilder values;
-    switch (variant.value_or(FontVariantEastAsianVariant::Normal)) {
-    case FontVariantEastAsianVariant::Normal:
-        break;
-    case FontVariantEastAsianVariant::Jis78:
-        values.append(CSSPrimitiveValue::create(CSSValueJis78));
-        break;
-    case FontVariantEastAsianVariant::Jis83:
-        values.append(CSSPrimitiveValue::create(CSSValueJis83));
-        break;
-    case FontVariantEastAsianVariant::Jis90:
-        values.append(CSSPrimitiveValue::create(CSSValueJis90));
-        break;
-    case FontVariantEastAsianVariant::Jis04:
-        values.append(CSSPrimitiveValue::create(CSSValueJis04));
-        break;
-    case FontVariantEastAsianVariant::Simplified:
-        values.append(CSSPrimitiveValue::create(CSSValueSimplified));
-        break;
-    case FontVariantEastAsianVariant::Traditional:
-        values.append(CSSPrimitiveValue::create(CSSValueTraditional));
-        break;
-    }
-
-    switch (width.value_or(FontVariantEastAsianWidth::Normal)) {
-    case FontVariantEastAsianWidth::Normal:
-        break;
-    case FontVariantEastAsianWidth::Full:
-        values.append(CSSPrimitiveValue::create(CSSValueFullWidth));
-        break;
-    case FontVariantEastAsianWidth::Proportional:
-        values.append(CSSPrimitiveValue::create(CSSValueProportionalWidth));
-        break;
-    }
-
-    switch (ruby.value_or(FontVariantEastAsianRuby::Normal)) {
-    case FontVariantEastAsianRuby::Normal:
-        break;
-    case FontVariantEastAsianRuby::Yes:
-        values.append(CSSPrimitiveValue::create(CSSValueRuby));
-    }
-
-    if (values.isEmpty())
-        return nullptr;
-
-    return CSSValueList::createSpaceSeparated(WTFMove(values));
-}
-
-RefPtr<CSSValue> consumeFontVariantAlternates(CSSParserTokenRange& range)
-{
-    if (range.atEnd())
-        return nullptr;
-
-    if (range.peek().id() == CSSValueNormal) {
-        consumeIdent<CSSValueNormal>(range);
-        return CSSPrimitiveValue::create(CSSValueNormal);
-    }
-
-    auto result = FontVariantAlternates::Normal();
-
-    auto parseSomethingWithoutError = [&range, &result]() {
-        bool hasParsedSomething = false;
-        auto parseAndSetArgument = [&range, &hasParsedSomething] (String& value) {
-            if (!value.isNull())
-                return false;
-
-            CSSParserTokenRange args = consumeFunction(range);
-            auto ident = consumeCustomIdent(args);
-            if (!args.atEnd())
-                return false;
-
-            if (!ident)
-                return false;
-
-            hasParsedSomething = true;
-            value = ident->stringValue();
-            return true;
-        };
-        auto parseListAndSetArgument = [&range, &hasParsedSomething] (Vector<String>& value) {
-            if (!value.isEmpty())
-                return false;
-
-            CSSParserTokenRange args = consumeFunction(range);
-            do {
-                auto ident = consumeCustomIdent(args);
-                if (!ident)
-                    return false;
-                value.append(ident->stringValue());
-            } while (consumeCommaIncludingWhitespace(args));
-
-            if (!args.atEnd())
-                return false;
-
-            hasParsedSomething = true;
-            return true;
-        };
-        while (true) {
-            const CSSParserToken& token = range.peek();
-            if (token.id() == CSSValueHistoricalForms) {
-                consumeIdent<CSSValueHistoricalForms>(range);
-
-                if (result.valuesRef().historicalForms)
-                    return false;
-
-                if (result.isNormal())
-                    result.setValues();
-
-                hasParsedSomething = true;
-                result.valuesRef().historicalForms = true;
-            } else if (token.functionId() == CSSValueSwash) {
-                if (!parseAndSetArgument(result.valuesRef().swash))
-                    return false;
-            } else if (token.functionId() == CSSValueStylistic) {
-                if (!parseAndSetArgument(result.valuesRef().stylistic))
-                    return false;
-            } else if (token.functionId() == CSSValueStyleset) {
-                if (!parseListAndSetArgument(result.valuesRef().styleset))
-                    return false;
-            } else if (token.functionId() == CSSValueCharacterVariant) {
-                if (!parseListAndSetArgument(result.valuesRef().characterVariant))
-                    return false;
-            } else if (token.functionId() == CSSValueOrnaments) {
-                if (!parseAndSetArgument(result.valuesRef().ornaments))
-                    return false;
-            } else if (token.functionId() == CSSValueAnnotation) {
-                if (!parseAndSetArgument(result.valuesRef().annotation))
-                    return false;
-            } else
-                return hasParsedSomething;
-        }
-    };
-
-    if (parseSomethingWithoutError())
-        return CSSFontVariantAlternatesValue::create(WTFMove(result));
-
-    return nullptr;
-}
-
-RefPtr<CSSValue> consumeFontVariantNumeric(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueNormal)
-        return consumeIdent(range);
-
-    CSSFontVariantNumericParser numericParser;
-    do {
-        if (numericParser.consumeNumeric(range) != CSSFontVariantNumericParser::ParseResult::ConsumedValue)
-            return nullptr;
-    } while (!range.atEnd());
-
-    return numericParser.finalizeValue();
-}
-
-RefPtr<CSSValue> consumeFontWeight(CSSParserTokenRange& range)
-{
-    if (auto result = consumeFontWeightRaw(range)) {
-        return WTF::switchOn(*result, [] (CSSValueID valueID) {
-            return CSSPrimitiveValue::create(valueID);
-        }, [] (double weightNumber) {
-            return CSSPrimitiveValue::create(weightNumber);
-        });
-    }
-    return nullptr;
-}
-
-RefPtr<CSSValue> consumeFamilyName(CSSParserTokenRange& range)
-{
-    auto familyName = consumeFamilyNameRaw(range);
-    if (familyName.isNull())
-        return nullptr;
-    return CSSValuePool::singleton().createFontFamilyValue(familyName);
-}
-
-RefPtr<CSSValue> consumeFamilyNameList(CSSParserTokenRange& range)
-{
-    return consumeCommaSeparatedListWithoutSingleValueOptimization(range, [] (auto& range) {
-        return consumeFamilyName(range);
-    });
-}
-
-RefPtr<CSSValue> consumeFontFamily(CSSParserTokenRange& range)
-{
-    return consumeCommaSeparatedListWithoutSingleValueOptimization(range, [] (auto& range) -> RefPtr<CSSValue> {
-        if (auto parsedValue = consumeGenericFamily(range))
-            return parsedValue;
-        return consumeFamilyName(range);
-    });
 }
 
 static RefPtr<CSSValue> consumeCounter(CSSParserTokenRange& range, int defaultValue)
@@ -4374,89 +3776,6 @@ RefPtr<CSSValue> consumeViewTransitionName(CSSParserTokenRange& range)
 }
 
 // MARK: - @-rule descriptor consumers:
-
-// MARK: @font-face
-
-RefPtr<CSSValue> consumeFontFaceFontFamily(CSSParserTokenRange& range)
-{
-    auto name = consumeFamilyName(range);
-    if (!name || !range.atEnd())
-        return nullptr;
-
-    // FIXME-NEWPARSER: https://bugs.webkit.org/show_bug.cgi?id=196381 For compatibility with the old parser, we have to make
-    // a list here, even though the list always contains only a single family name.
-    // Once the old parser is gone, we can delete this function, make the caller
-    // use consumeFamilyName instead, and then patch the @font-face code to
-    // not expect a list with a single name in it.
-    return CSSValueList::createCommaSeparated(name.releaseNonNull());
-}
-
-Vector<FontTechnology> consumeFontTech(CSSParserTokenRange& range, bool singleValue)
-{
-    Vector<FontTechnology> technologies;
-    auto args = consumeFunction(range);
-    do {
-        auto& arg = args.consumeIncludingWhitespace();
-        if (arg.type() != IdentToken)
-            return { };
-        auto technology = fromCSSValueID<FontTechnology>(arg.id());
-        if (technology == FontTechnology::Invalid)
-            return { };
-        technologies.append(technology);
-    } while (consumeCommaIncludingWhitespace(args) && !singleValue);
-    if (!args.atEnd())
-        return { };
-    return technologies;
-}
-
-static bool isFontFormatKeywordValid(CSSValueID id)
-{
-    switch (id) {
-    case CSSValueCollection:
-    case CSSValueEmbeddedOpentype:
-    case CSSValueOpentype:
-    case CSSValueSvg:
-    case CSSValueTruetype:
-    case CSSValueWoff:
-    case CSSValueWoff2:
-        return true;
-    default:
-        return false;
-    }
-}
-
-String consumeFontFormat(CSSParserTokenRange& range, bool rejectStringValues)
-{
-    // https://drafts.csswg.org/css-fonts/#descdef-font-face-src
-    // FIXME: We allow any identifier here and convert to strings; specification calls for certain keywords and legacy compatibility strings.
-    auto args = CSSPropertyParserHelpers::consumeFunction(range);
-    auto& arg = args.consumeIncludingWhitespace();
-    if (!args.atEnd())
-        return nullString();
-    if (arg.type() == IdentToken && isFontFormatKeywordValid(arg.id()))
-        return arg.value().toString();
-    if (arg.type() == StringToken && !rejectStringValues)
-        return arg.value().toString();
-    return nullString();
-}
-
-// MARK: @font-palette-values
-
-RefPtr<CSSValue> consumeFontPaletteValuesOverrideColors(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    auto list = consumeCommaSeparatedListWithoutSingleValueOptimization(range, [](auto& range, auto& context) -> RefPtr<CSSValue> {
-        auto key = consumeNonNegativeInteger(range);
-        if (!key)
-            return nullptr;
-        auto color = consumeColor(range, context, { .allowedColorTypes = StyleColor::CSSColorType::Absolute });
-        if (!color)
-            return nullptr;
-        return CSSFontPaletteValuesOverrideColorsValue::create(key.releaseNonNull(), color.releaseNonNull());
-    }, context);
-    if (!range.atEnd() || !list || !list->length())
-        return nullptr;
-    return list;
-}
 
 // MARK: @counter-style
