+++ Source/WebCore/style/StyleResolveForFontRaw.cpp
@@ -33,6 +33,7 @@
 
 #include "CSSFontSelector.h"
 #include "CSSHelper.h"
+#include "CSSPropertyParserConsumer+Font.h"
 #include "CSSPropertyParserHelpers.h"
 #include "Document.h"
 #include "DocumentInlines.h"
